Create README.md
Windshield wipers and car batteries are consumables for cars, generally, the life of windshield wipers is one year, and the life of the car batteries is 2 to 3 years. Forecasting sales is considered to be one of the most important strategies of a car accessories company as it can decide how much goods to supply to customers in the next period. Finding factors that affect sales is crucial to forecast the next period of sales. In this study, sales data from CarQuality in 2010-2011, consisting of area, date, time, commodity category, ID and sales, weather data from a government website in 2010-2011, consisting of daily rain precipitation, air pressure, temperature, humidity, and water pressure, are used to develop forecasting models. Three different models including weather only, weather plus time lag, and the five most important features are investigated and then compared by their respective performance to forecast the sales amount for wiper and battery in 2011. The result showed signs of overfitting, we conclude that more data is needed to capture the yearly trend for the sales of windshield wipers and batteries.